+++ hw-ledger-transports/node.ts
@@ -1,5 +1,5 @@
 
-import LedgerHid from 'https://esm.sh/@ledgerhq/hw-transport-node-hid-singleton@6.28.18';
+import LedgerHid from 'https://esm.sh/@ledgerhq/hw-transport-node-hid-singleton@6.29.0';
 
 import { createDefs } from './util.ts';