+++ util-crypto/address/eq.ts
@@ -1,5 +1,5 @@
 
-import { u8aEq } from 'https://deno.land/x/polkadot@0.2.43/util/mod.ts';
+import { u8aEq } from 'https://deno.land/x/polkadot/util/mod.ts';
 
 import { decodeAddress } from './decode.ts';
 
@@ -12,7 +12,7 @@ import { decodeAddress } from './decode.ts';
  * <BR>
  *
  * ```javascript
- * import { u8aEq } from 'https://deno.land/x/polkadot@0.2.43/util/mod.ts';
+ * import { u8aEq } from 'https://deno.land/x/polkadot/util/mod.ts';
  *
  * u8aEq(new Uint8Array([0x68, 0x65]), new Uint8Array([0x68, 0x65])); // true
  * ```