+++ util-crypto/ed25519/verify.ts
@@ -1,8 +1,8 @@
 
 import { ed25519 } from 'https://esm.sh/@noble/curves@1.2.0/ed25519.js';
 
-import { hasBigInt, u8aToU8a } from 'https://deno.land/x/polkadot@0.2.43/util/mod.ts';
-import { ed25519Verify as wasmVerify, isReady } from 'https://deno.land/x/polkadot@0.2.43/wasm-crypto/mod.ts';
+import { hasBigInt, u8aToU8a } from 'https://deno.land/x/polkadot/util/mod.ts';
+import { ed25519Verify as wasmVerify, isReady } from 'https://deno.land/x/polkadot/wasm-crypto/mod.ts';
 
 /**
  * @name ed25519Sign
@@ -13,7 +13,7 @@ import { ed25519Verify as wasmVerify, isReady } from 'https://deno.land/x/polkad
  * <BR>
  *
  * ```javascript
- * import { ed25519Verify } from 'https://deno.land/x/polkadot@0.2.43/util-crypto/mod.ts';
+ * import { ed25519Verify } from 'https://deno.land/x/polkadot/util-crypto/mod.ts';
  *
  * ed25519Verify([...], [...], [...]); // => true/false
  * ```