+++ util-crypto/nacl/encrypt.ts
@@ -16,7 +16,7 @@ interface Encrypted {
  * <BR>
  *
  * ```javascript
- * import { naclEncrypt } from 'https://deno.land/x/polkadot@0.2.43/util-crypto/mod.ts';
+ * import { naclEncrypt } from 'https://deno.land/x/polkadot/util-crypto/mod.ts';
  *
  * naclEncrypt([...], [...]); // => [...]
  * ```