+++ util-crypto/sr25519/agreement.ts
@@ -1,6 +1,6 @@
 
-import { u8aToU8a } from 'https://deno.land/x/polkadot@0.2.43/util/mod.ts';
-import { sr25519Agree } from 'https://deno.land/x/polkadot@0.2.43/wasm-crypto/mod.ts';
+import { u8aToU8a } from 'https://deno.land/x/polkadot/util/mod.ts';
+import { sr25519Agree } from 'https://deno.land/x/polkadot/wasm-crypto/mod.ts';
 
 /**
  * @name sr25519Agreement