+++ x-ws/shim.ts
@@ -1,5 +1,5 @@
 
-import { exposeGlobal } from 'https://deno.land/x/polkadot@0.2.43/x-global/mod.ts';
-import { WebSocket } from 'https://deno.land/x/polkadot@0.2.43/x-ws/mod.ts';
+import { exposeGlobal } from 'https://deno.land/x/polkadot/x-global/mod.ts';
+import { WebSocket } from 'https://deno.land/x/polkadot/x-ws/mod.ts';
 
 exposeGlobal('WebSocket', WebSocket);